refactor(api): Reorganize API-related components
Created api-hooks, game-api, and game-config; removed api and game
