+++ src/game-instance.ts
