--- src/game.ts
@@ -1,25 +1,0 @@
-import GameInstance from "./gameInstance";
-import { RegalError } from "./error";
-
-export class Game {
-
-    private static _onGameStart = (): GameInstance => {
-        throw new RegalError("onGameStart has not been implemented by the game developer.")
-    }
-    static get onGameStart() {
-        return this._onGameStart;
-    }
-    static set onGameStart(startFunc: () => GameInstance) {
-        this._onGameStart = startFunc;
-    }
-
-    private static _onUserInput = (content: string, game: GameInstance): GameInstance => {
-        throw new RegalError("onUserInput has not been implemented by the game developer.")
-    }
-    static get onUserInput() {
-        return this._onUserInput;
-    }
-    static set onUserInput(inputFunc: (content: string, game: GameInstance) => GameInstance) {
-        this._onUserInput = inputFunc;
-    }
-}
