+++ src/index.ts
@@ -1,6 +1,8 @@
-import * as Api from './api';
-import * as Event from './event';
-
-export { Api };
-export { Event };
-export * from './game';
+export * from './agent'; // TODO - Decide what to export from agent.ts
+export { onPlayerCommand, onStartCommand, onBeforeUndoCommand } from './api-hooks';
+export { ErrorCode, RegalError } from './error';
+export { EventFunction, TrackedEvent, EventQueue, noop, EventRecord, InstanceEvents, enqueue, nq, on } from './event';
+export { Game, GameResponse } from './game-api';
+export { GameOptions, GameMetadata } from './game-config';
+export * from './game-instance';
+export { OutputLineType, OutputLine, InstanceOutput, GameOutput } from './output';
