+++ test/output.test.ts
@@ -1,7 +1,7 @@
 import { expect } from 'chai';
 import 'mocha';
 
-import GameInstance from '../src/gameInstance';
+import GameInstance from '../src/game-instance';
 import { OutputLineType } from '../src/output';
 
 describe("Output", function() {
